# dict: use with_capacity in HashMap extraction