--- a/benches/bench_dict.rs
+++ b/benches/bench_dict.rs
@@ -3,6 +3,7 @@
 extern crate test;
 use pyo3::prelude::*;
 use pyo3::types::IntoPyDict;
+use std::collections::{BTreeMap, HashMap};
 use test::Bencher;
 
 #[bench]
@@ -32,4 +33,32 @@ fn dict_get_item(b: &mut Bencher) {
             sum += dict.get_item(i).unwrap().extract::<usize>().unwrap();
         }
     });
+}
+
+#[bench]
+fn extract_hashmap(b: &mut Bencher) {
+    let gil = Python::acquire_gil();
+    let py = gil.python();
+    const LEN: usize = 100_000;
+    let dict = (0..LEN as u64).map(|i| (i, i * 2)).into_py_dict(py);
+    b.iter(|| HashMap::<u64, u64>::extract(dict));
+}
+
+#[bench]
+fn extract_btreemap(b: &mut Bencher) {
+    let gil = Python::acquire_gil();
+    let py = gil.python();
+    const LEN: usize = 100_000;
+    let dict = (0..LEN as u64).map(|i| (i, i * 2)).into_py_dict(py);
+    b.iter(|| BTreeMap::<u64, u64>::extract(dict));
+}
+
+#[bench]
+#[cfg(feature = "hashbrown")]
+fn extract_hashbrown_map(b: &mut Bencher) {
+    let gil = Python::acquire_gil();
+    let py = gil.python();
+    const LEN: usize = 100_000;
+    let dict = (0..LEN as u64).map(|i| (i, i * 2)).into_py_dict(py);
+    b.iter(|| hashbrown::HashMap::<u64, u64>::extract(dict));
 }