A tiny PR to improve performance when extracting `HashMap` from Python.

I added some new benchmarks to prove it was worth adding. For the two relevant benchmarks, I get the following results.

Before:

```
test extract_hashbrown_map ... bench:   4,310,640 ns/iter (+/- 438,993)
test extract_hashmap       ... bench:   6,350,030 ns/iter (+/- 499,598)
```

After:

```
test extract_hashbrown_map ... bench:   3,645,450 ns/iter (+/- 408,332)
test extract_hashmap       ... bench:   4,759,320 ns/iter (+/- 285,136)
```

So definitely helps.